Move main function into separate package (closes #12)
